Feat : seperate cacluate num file
