--- src/routes/Dashboard/CardList/utils.ts
@@ -1,45 +1,0 @@
-interface IProps {
-  category: string
-  value: number[]
-  prevValue: number[]
-}
-
-const dataStructure: IProps[] = [
-  { category: 'ROAS', value: [], prevValue: [] },
-  { category: '광고비', value: [], prevValue: [] },
-  { category: '노출수', value: [], prevValue: [] },
-  { category: '클릭수', value: [], prevValue: [] },
-  { category: '전환수', value: [], prevValue: [] },
-  { category: '매출', value: [], prevValue: [] },
-]
-
-export const translateData = (DATA: ITrendData[], PAST_DATA: ITrendData[]) => {
-  const data = dataStructure
-  DATA.forEach((d) => {
-    data.find((item) => item.category === 'ROAS')!.value.push(d.roas)
-    data.find((item) => item.category === '광고비')!.value.push(d.cost)
-    data.find((item) => item.category === '노출수')!.value.push(d.imp)
-    data.find((item) => item.category === '클릭수')!.value.push(d.click)
-    data.find((item) => item.category === '전환수')!.value.push(d.conv)
-    data.find((item) => item.category === '매출')!.value.push(d.convValue)
-  })
-
-  PAST_DATA.forEach((d) => {
-    data.find((item) => item.category === 'ROAS')!.prevValue.push(d.roas)
-    data.find((item) => item.category === '광고비')!.prevValue.push(d.cost)
-    data.find((item) => item.category === '노출수')!.prevValue.push(d.imp)
-    data.find((item) => item.category === '클릭수')!.prevValue.push(d.click)
-    data.find((item) => item.category === '전환수')!.prevValue.push(d.conv)
-    data.find((item) => item.category === '매출')!.prevValue.push(d.convValue)
-  })
-
-  return data
-}
-
-export const transformNum = (NUM: number, TYPE: string) => {
-  if (TYPE === 'ROAS') return { unitCnt: Math.floor(NUM), unitWord: '' }
-  const unitCnt = Math.ceil(NUM / 1000)
-  const unit = ['', '천', '만', '억', '조']
-  const unitWord = unit[unitCnt.toString().length]
-  return { unitCnt, unitWord }
-}
